damage-detection/requirements.txt
keras==2.8.0
matplotlib==3.2.2
numpy==1.21.6
opencv-contrib-python==3.4.13.47
pandas==1.3.5
Pillow==9.1.1
scikit_image==0.18.3
scikit_learn==1.1.1
seaborn==0.11.2
streamlit==1.9.2
tensorflow==2.8.0
